Moving http/grpc specifc methods from tracing to http/grpc_tracing files.
