+++ pkg/diagnostics/tracing.go
@@ -12,17 +12,12 @@ import (
 	"encoding/hex"
 	"fmt"
 	"math/rand"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 
 	"github.com/dapr/dapr/pkg/config"
 	diag_utils "github.com/dapr/dapr/pkg/diagnostics/utils"
-	invokev1 "github.com/dapr/dapr/pkg/messaging/v1"
-	internalv1pb "github.com/dapr/dapr/pkg/proto/internals/v1"
-	runtimev1pb "github.com/dapr/dapr/pkg/proto/runtime/v1"
-	"github.com/valyala/fasthttp"
 
 	"go.opencensus.io/trace"
 	"google.golang.org/grpc/metadata"
@@ -202,41 +197,6 @@ func extractDaprMetadata(ctx context.Context) map[string][]string {
 	return daprMetadata
 }
 
-func getSpanAttributesMapFromHTTPContext(ctx *fasthttp.RequestCtx) map[string]string {
-	// Span Attribute reference https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/trace/semantic_conventions
-	route := string(ctx.Request.URI().Path())
-	method := string(ctx.Request.Header.Method())
-	uri := ctx.Request.URI().String()
-	statusCode := ctx.Response.StatusCode()
-	r := getAPIComponent(route)
-	return GetSpanAttributesMapFromHTTP(r.componentType, r.componentValue, method, route, uri, statusCode)
-}
-
-// GetSpanAttributesMap builds the span trace attributes map for HTTP calls based on given parameters as per open-telemetry specs
-func GetSpanAttributesMapFromHTTP(componentType, componentValue, method, route, uri string, statusCode int) map[string]string {
-	// Span Attribute reference https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/trace/semantic_conventions
-	m := make(map[string]string)
-	switch componentType {
-	case "state", "secrets", "bindings":
-		m[dbTypeSpanAttributeKey] = componentType
-		m[dbInstanceSpanAttributeKey] = componentValue
-		// TODO: not possible currently to get the route {state_store} , so using path instead of route
-		m[dbStatementSpanAttributeKey] = fmt.Sprintf("%s %s", method, route)
-		m[dbURLSpanAttributeKey] = route
-	case "invoke", "actors":
-		m[httpMethodSpanAttributeKey] = method
-		m[httpURLSpanAttributeKey] = uri
-		code := invokev1.CodeFromHTTPStatus(statusCode)
-		m[httpStatusCodeSpanAttributeKey] = strconv.Itoa(statusCode)
-		m[httpStatusTextSpanAttributeKey] = code.String()
-	case "publish":
-		m[messagingSystemSpanAttributeKey] = componentType
-		m[messagingDestinationSpanAttributeKey] = componentValue
-		m[messagingDestinationKindSpanAttributeKey] = messagingDestinationKind
-	}
-	return m
-}
-
 // AddAttributesToSpan adds the given attributes in the span
 func AddAttributesToSpan(span *trace.Span, attributes map[string]string) {
 	if span != nil {
@@ -268,83 +228,6 @@ func getAPIComponent(apiPath string) apiComponent {
 
 	// return 'state', 'statestore' from the parsed tokens in apiComponent type
 	return apiComponent{componentType: tokens[1], componentValue: tokens[2]}
-}
-
-// GetSpanAttributesMapFromGRPC builds the span trace attributes map for gRPC calls based on given parameters as per open-telemetry specs
-func GetSpanAttributesMapFromGRPC(req interface{}, rpcMethod string) map[string]string {
-	// RPC Span Attribute reference https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/rpc.md
-	// gRPC method /package.service/method
-	var serviceName string
-
-	p := rpcMethod
-	if p[0] == '/' {
-		p = rpcMethod[1:]
-	}
-
-	// Split up to 3 delimiters in '/package.service/method'
-	// example for internal call : "/dapr.proto.internals.v1.ServiceInvocation/CallLocal"
-	// example for non-internal call : "/dapr.proto.runtime.v1.Dapr/InvokeService"
-	internalCall := strings.Contains(p, "dapr.proto.internals.")
-
-	tokens := strings.SplitN(p, "/", 3)
-	if len(tokens) >= 2 {
-		if internalCall {
-			if t := strings.SplitN(tokens[0], ".", 5); len(t) >= 4 {
-				serviceName = t[4]
-			}
-		} else {
-			// TODO : need to revisit /package.service/method format in runtime calls, currently service name is "Dapr"
-			// so instead of taking general "Dapr", taking method name as service name to give better insights
-			serviceName = tokens[1]
-		}
-	} else {
-		// default service name to full method name, after removing first "/"
-		serviceName = p
-	}
-
-	m := make(map[string]string)
-	m[gRPCServiceSpanAttributeKey] = serviceName
-
-	// below attribute is not as per open temetery spec, adding custom dapr attribute to have additional details
-	m[gRPCDaprInstanceSpanAttributeKey] = extractComponentValueFromGRPCRequest(req)
-	return m
-}
-
-func extractComponentValueFromGRPCRequest(req interface{}) string {
-	if req == nil {
-		return ""
-	}
-	if s, ok := req.(*internalv1pb.InternalInvokeRequest); ok {
-		return s.Message.GetMethod()
-	}
-	if s, ok := req.(*runtimev1pb.PublishEventRequest); ok {
-		return s.GetTopic()
-	}
-	if s, ok := req.(*runtimev1pb.InvokeServiceRequest); ok {
-		return s.Message.GetMethod()
-	}
-	if s, ok := req.(*runtimev1pb.InvokeBindingRequest); ok {
-		return s.GetName()
-	}
-	if s, ok := req.(*runtimev1pb.GetStateRequest); ok {
-		return s.GetStoreName()
-	}
-	if s, ok := req.(*runtimev1pb.GetSecretRequest); ok {
-		return s.GetStoreName()
-	}
-	if s, ok := req.(*runtimev1pb.SaveStateRequest); ok {
-		return s.GetStoreName()
-	}
-	if s, ok := req.(*runtimev1pb.DeleteStateRequest); ok {
-		return s.GetStoreName()
-	}
-	if s, ok := req.(*runtimev1pb.TopicEventRequest); ok {
-		return s.GetTopic()
-	}
-	if s, ok := req.(*runtimev1pb.BindingEventRequest); ok {
-		return s.GetName()
-	}
-	return ""
 }
 
 var tracingConfig atomic.Value // access atomically
